requirements.txt
apache-airflow==2.4.0
pandas==2.2.0
apache-airflow-providers-snowflake==4.4.2
mysql-connector-python==8.0.31
pymysql